requirements.txt
PyInquirer
prompt_toolkit==1.0.14
travispy
gitpython
pyyaml
setuptools